@@ -4,6 +4,7 @@
   "private": true,
   "description": "Личный проект «Кекстаграм» от HTML Academy",
   "author": "Nikolay Solodukhin",
+  "publicUrl": "",
   "repository": {
     "type": "git",
     "url": "git@github.com:NikolaySolodukhin/kekstagram.git"
@@ -12,23 +13,52 @@
     "url": "https://github.com/NikolaySolodukhin/kekstagram/issues"
   },
   "devDependencies": {
-    "colors": "1.1.2",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.2.3",
+    "babel-loader": "^7.1.4",
+    "babel-plugin-syntax-dynamic-import": "^6.18.0",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
+    "caniuse-lite": "^1.0.30000830",
+    "case-sensitive-paths-webpack-plugin": "^2.1.1",
     "copy-webpack-plugin": "^3.0.1",
-    "eslint": "3.2.0",
+    "css-loader": "^0.28.11",
+    "csso-webpack-plugin": "^1.0.0-beta.12",
+    "eslint": "^4.15.0",
+    "eslint-plugin-prettier": "^2.6.0",
+    "eslint-plugin-react": "^7.7.0",
     "eslint-plugin-requirejs": "2.0.1",
-    "express": "4.14.0",
-    "finalhandler": "0.5.0",
-    "http-server": "0.9.0",
-    "imagemagick": "0.1.3",
-    "log4js": "0.6.37",
-    "opener": "1.4.1",
-    "webpack": "1.13.2",
-    "webpack-dev-middleware": "1.6.1"
+    "extract-text-webpack-plugin": "^3.0.2",
+    "file-loader": "^1.1.11",
+    "html-webpack-plugin": "^3.2.0",
+    "postcss-cssnext": "^3.1.0",
+    "postcss-import": "^11.1.0",
+    "postcss-loader": "^2.1.4",
+    "postcss-nested": "^3.0.0",
+    "postcss-url": "^7.3.2",
+    "prettier": "^1.12.1",
+    "raw-loader": "^0.5.1",
+    "react-dev-utils": "^5.0.1",
+    "style-loader": "^0.21.0",
+    "webpack-bundle-analyzer": "^2.11.1",
+    "webpack-chunk-hash": "^0.6.0",
+    "webpack-dev-middleware": "1.6.1",
+    "webpack-dev-server": "^2.11.0",
+    "webpack-manifest-plugin": "^2.0.1"
   },
-  "dependencies": {},
+  "dependencies": {
+    "classlist-polyfill": "^1.2.0",
+    "core-js": "^2.5.5",
+    "cross-env": "^5.1.4",
+    "webpack": "^3.10.0",
+    "whatwg-fetch": "^2.0.4"
+  },
   "scripts": {
-    "build": "webpack --config ./bin/webpack.config.js",
-    "start": "node ./bin/server.js kekstagram"
+    "clean": "rm -rf ./build",
+    "build": "npm run clean && cross-env NODE_ENV=production webpack",
+    "start": "cross-env NODE_ENV=development webpack-dev-server --color --progress",
+    "analize-bundle": "cross-env NODE_ENV=production webpack --profile --json > stats.json && webpack-bundle-analyzer stats.json"
   },
   "engines": {
     "node": "^5.11.0"
