@@ -28,9 +28,11 @@ Gallery.prototype.show = function(number) {
   this.overlay.classList.remove('invisible');
   this.setActivePicture(number);
 };
+
 Gallery.prototype.onGalleryImageClick = function() {
   this.setActivePicture((this.activePicture + 1) % this.allPictures.length);
 };
+
 Gallery.prototype.hide = function() {
   this.overlay.classList.add('invisible');
   this.overlayClose.addEventListener('click', null);
@@ -39,9 +41,13 @@ Gallery.prototype.hide = function() {
 
 Gallery.prototype.setActivePicture = function(number) {
   this.activePicture = number;
-  this.overlayImage.src = this.allPictures[number].url;
-  this.overlay.querySelector('.likes-count').textContent = this.allPictures[number].likes;
-  this.overlay.querySelector('.comments-count').textContent = this.allPictures[number].comments;
+  this.overlayImage.src = require(`../assets/photos/${number + 1}.jpg`);
+  this.overlay.querySelector('.likes-count').textContent = this.allPictures[
+    number
+  ].likes;
+  this.overlay.querySelector('.comments-count').textContent = this.allPictures[
+    number
+  ].comments;
 };
 
-module.exports = new Gallery();
+export default new Gallery();
