@@ -1,55 +1,59 @@
 'use strict';
 
-var load = require('./load');
-var Picture = require('./picture');
-var gallery = require('./gallery');
-var throttle = require('./throttle');
+import load from './load';
+import Picture from './picture';
+import gallery from './gallery';
+import throttle from './throttle';
 
-var filters = document.querySelector('.filters');
-var footer = document.querySelector('.footer');
-var picturesContainer = document.querySelector('.pictures');
-var PICTURE_LOAD_URL = '/api/pictures';
-
-/** @constant {number} */
-var PAGE_SIZE = 12;
+const filters = document.querySelector('.filters');
+const footer = document.querySelector('.footer');
+const picturesContainer = document.querySelector('.pictures');
+const PICTURE_LOAD_URL = '/data/data.json';
 
 /** @constant {number} */
-var THROTTLE_TIMEOUT = 100;
+const PAGE_SIZE = 12;
 
 /** @constant {number} */
-var GAP = 180;
+const THROTTLE_TIMEOUT = 100;
+
+/** @constant {number} */
+const GAP = 180;
 
 /** @constant {string} */
-var DEFAULT_FILTER = 'filter-popular';
+const DEFAULT_FILTER = 'filter-popular';
 
 /** @type {string} */
-var activeFilter = localStorage.getItem('filter') || DEFAULT_FILTER;
+let activeFilter = localStorage.getItem('filter') || DEFAULT_FILTER;
 
 /** @type {number} */
-var pageNumber = 0;
+let pageNumber = 0;
 
 filters.classList.add('hidden');
 
-var loadPictures = function(filter, currentPage) {
-  load(PICTURE_LOAD_URL, {
-    from: currentPage * PAGE_SIZE,
-    to: currentPage * PAGE_SIZE + PAGE_SIZE,
-    filter: filter
-  }, function(pictures) {
-    showPictures(pictures);
-  });
+const loadPictures = function(filter, currentPage) {
+  load(
+    PICTURE_LOAD_URL,
+    {
+      from: currentPage * PAGE_SIZE,
+      to: currentPage * PAGE_SIZE + PAGE_SIZE,
+      filter: filter,
+    },
+    pictures => showPictures(pictures)
+  );
 };
+
 /**
  * Принимает на вход массив с элементами к каждой фотографии/ Записывает все фотографии в один контейнер DocumentFragment и затем добавляет их в DOM
  * @param {object[]} picturesAll
  */
 
 function showPictures(picturesAll) {
+  const picturesContainerAll = document.createDocumentFragment();
 
-  var picturesContainerAll = document.createDocumentFragment();
-
-  picturesAll.forEach(function(pictureData, number) {
-    picturesContainerAll.appendChild(new Picture(pictureData, pageNumber * PAGE_SIZE + number).element);
+  picturesAll.forEach((pictureData, number) => {
+    picturesContainerAll.appendChild(
+      new Picture(pictureData, pageNumber * PAGE_SIZE + number).element
+    );
   });
 
   picturesContainer.appendChild(picturesContainerAll);
@@ -59,28 +63,36 @@ function showPictures(picturesAll) {
 }
 
 /**  Обработчик кликов по фильтрам */
-var setFiltersEnabled = function() {
-  filters.addEventListener('change', function(event) {
-    if (event.target.name === 'filter') {
+const setFiltersEnabled = () => {
+  filters.addEventListener(
+    'change',
+    event => {
+      if (event.target.name !== 'filter') {
+        return;
+      }
       gallery.clearPictures();
       picturesContainer.innerHTML = '';
       pageNumber = 0;
       activeFilter = event.target.id;
       localStorage.setItem('filter', activeFilter);
       loadPictures(activeFilter, pageNumber);
-    }
-  }, true);
+    },
+    true
+  );
 };
 
-var isFooterVisible = function() {
-  return (footer.getBoundingClientRect().top - window.innerHeight) <= GAP;
-};
+const isFooterVisible = () =>
+  footer.getBoundingClientRect().top - window.innerHeight <= GAP;
+
 /** Обработчик прокрутки */
-window.addEventListener('scroll', throttle(function() {
-  if (isFooterVisible()) {
-    loadPictures(activeFilter, ++pageNumber);
-  }
-}, THROTTLE_TIMEOUT));
+window.addEventListener(
+  'scroll',
+  throttle(() => {
+    if (isFooterVisible()) {
+      loadPictures(activeFilter, ++pageNumber);
+    }
+  }, THROTTLE_TIMEOUT)
+);
 
 loadPictures(DEFAULT_FILTER, pageNumber);
 setFiltersEnabled();
